fix BlockExecution spec
